Added Total and tax to order

+++ Turnup/Services/ProductService/IProductService.cs
@@ -6,5 +6,5 @@ public interface IProductService
 {
     Task<ServiceResponse<List<Product>>> GetProductsAsync(string establishmentId);
 
-    Task<ServiceResponse<Product>> CreateNewProduct(string title, string description, string imageUrl, long price, string establishmentId);
+    Task<ServiceResponse<Product>> CreateNewProduct(string title, string description, string imageUrl, decimal price, string establishmentId);
 }
